+++ src/init.lua
@@ -1,6 +1,7 @@
 local DoublyLinkedList = require(script.DoublyLinkedList)
 local LinkedList = require(script.LinkedList)
 local Map = require(script.Map)
+local OrderedSet = require(script.OrderedSet)
 local Queue = require(script.Queue)
 local Set = require(script.Set)
 local Stack = require(script.Stack)
@@ -9,6 +10,7 @@ return {
 	DoublyLinkedList = DoublyLinkedList,
 	LinkedList = LinkedList,
 	Map = Map,
+	OrderedSet = OrderedSet,
 	Queue = Queue,
 	Set = Set,
 	Stack = Stack,