+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Structures Changelog
 
 ## Unreleased Changes
+* Added the OrderedSet data structure.
 
 ## [v1.0.0](https://github.com/Nezuo/structures/releases/tag/v1.0.0) (December 2, 2020)
 * Added the DoublyLinkedList data structure.